requirements.txt
python-docx
Pillow
PyMuPDF
fpdf
pyttsx3
pyspellchecker